+++ tests/back/Pim/Enrichment/Integration/Storage/Sql/Completeness/SqlGetCompletenessProductMasksIntegration.php
@@ -7,31 +7,33 @@ namespace AkeneoTest\Pim\Enrichment\Integration\Storage\Sql\Completeness;
 use Akeneo\Pim\Enrichment\Bundle\Storage\Sql\Completeness\SqlGetCompletenessProductMasks;
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\CompletenessProductMask;
 use Akeneo\Pim\Enrichment\Component\Product\Model\PriceCollection;
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductPrice;
 use Akeneo\Pim\Enrichment\Component\Product\Model\WriteValueCollection;
 use Akeneo\Pim\Enrichment\Component\Product\Value\PriceCollectionValue;
 use Akeneo\Pim\Enrichment\Component\Product\Value\ScalarValue;
 use Akeneo\Test\Integration\TestCase;
+use Ramsey\Uuid\Uuid;
 
 class SqlGetCompletenessProductMasksIntegration extends TestCase
 {
     public function test_it_returns_mask_with_only_sku()
     {
-        $this->createProduct('simple_product', 'familyA', []);
+        $product = $this->createProduct('simple_product', 'familyA', []);
 
         $expected = [
             new CompletenessProductMask('-1', 'simple_product', 'familyA', [
                 'sku-<all_channels>-<all_locales>',
             ])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['simple_product']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_several_masks()
     {
-        $this->createProduct('product1', 'familyA', []);
-        $this->createProduct('product2', 'familyA', []);
+        $product1 = $this->createProduct('product1', 'familyA', []);
+        $product2 = $this->createProduct('product2', 'familyA', []);
 
         $expected = [
             new CompletenessProductMask('-1', 'product1', 'familyA', [
@@ -42,24 +44,24 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
             ])
         ];
 
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['product1', 'product2', 'nonExistingProduct']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product1->getUuid(), $product2->getUuid(), Uuid::uuid4()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_mask_for_a_product_without_family()
     {
-        $this->createProduct('product_without_family', null, []);
+        $product = $this->createProduct('product_without_family', null, []);
 
         $expected = [
             new CompletenessProductMask('-1', 'product_without_family', null, ['sku-<all_channels>-<all_locales>'])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['product_without_family']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_default_masks()
     {
-        $this->createProduct('complex_product', 'familyA', [
+        $product = $this->createProduct('complex_product', 'familyA', [
             'a_date' => [['locale' => null, 'scope' => null, 'data' => '2010-10-10']],
             'a_file' => [['locale' => null, 'scope' => null, 'data' => $this->getFileInfoKey($this->getFixturePath('akeneo.txt'))]],
             'a_metric' => [['locale' => null, 'scope' => null, 'data' => ['amount' => 1, 'unit' => 'WATT']]],
@@ -95,13 +97,13 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 'sku-<all_channels>-<all_locales>',
             ])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['complex_product']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_mask_with_specific_channels()
     {
-        $this->createProduct('product_with_scopable_data', 'familyA', [
+        $product = $this->createProduct('product_with_scopable_data', 'familyA', [
             'a_localized_and_scopable_text_area' => [
                 ['locale' => 'en_US', 'scope' => 'ecommerce', 'data' => 'foo']
             ],
@@ -117,13 +119,13 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 'a_scopable_price-USD-ecommerce-<all_locales>',
             ])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['product_with_scopable_data']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_mask_with_specific_locales()
     {
-        $this->createProduct('product_with_localizable_data', 'familyA', [
+        $product = $this->createProduct('product_with_localizable_data', 'familyA', [
             'a_localizable_image' => [
                 ['locale' => 'en_US', 'scope' => null, 'data' => $this->getFileInfoKey($this->getFixturePath('akeneo.jpg'))]
             ],
@@ -139,13 +141,13 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 'a_localized_and_scopable_text_area-ecommerce-en_US',
             ])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['product_with_localizable_data']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     public function test_it_returns_price_masks()
     {
-        $this->createProduct('product_with_prices', 'familyA', [
+        $product = $this->createProduct('product_with_prices', 'familyA', [
             'a_price' => [['locale' => null, 'scope' => null, 'data' => [['amount' => 50.00, 'currency' => 'EUR']]]],
             'a_scopable_price' => [
                 ['locale' => null, 'scope' => 'ecommerce', 'data' => [['amount' => '2000.00', 'currency' => 'USD']]]
@@ -159,13 +161,13 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 'a_scopable_price-USD-ecommerce-<all_locales>',
             ])
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['product_with_prices']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
     function test_that_it_returns_masks_even_if_an_attribute_was_deleted()
     {
-        $this->createProduct(
+        $product = $this->createProduct(
             'productA',
             'familyA',
             [
@@ -186,7 +188,7 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 ]
             )
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['productA']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
 
         $aPrice = $this->get('pim_catalog.repository.attribute')->findOneByIdentifier('a_price');
@@ -199,7 +201,7 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
                 ]
             ),
         ];
-        $result = $this->getCompletenessProductMasks()->fromProductIdentifiers(['productA']);
+        $result = $this->getCompletenessProductMasks()->fromProductUuids([$product->getUuid()]);
         $this->assertSameCompletenessProductMasks($expected, $result);
     }
 
@@ -250,7 +252,7 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
         return $this->catalog->useTechnicalSqlCatalog();
     }
 
-    private function createProduct(string $identifier, ?string $familyCode, array $values): void
+    private function createProduct(string $identifier, ?string $familyCode, array $values): ProductInterface
     {
         $product = $this->get('pim_catalog.builder.product')->createProduct($identifier, $familyCode);
         $this->get('pim_catalog.updater.product')->update($product, ['values' => $values]);
@@ -259,6 +261,8 @@ class SqlGetCompletenessProductMasksIntegration extends TestCase
             throw new \Exception(sprintf('Impossible to setup test in %s: %s', static::class, $errors->get(0)->getMessage()));
         }
         $this->get('pim_catalog.saver.product')->save($product);
+
+        return $this->get('pim_catalog.repository.product')->findOneByIdentifier($identifier);
     }
 
     private function getCompletenessProductMasks(): SqlGetCompletenessProductMasks {